% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindex.R
\name{reindex}
\alias{reindex}
\title{Reindex}
\usage{
reindex(x, time, base)
}
\arguments{
\item{x}{input data set.}

\item{time}{time variable}

\item{base}{new base. Must be date format.}
}
\description{
This function is intended to change the reference date of an index variable.
}
\examples{
library(lubridate)
x <- data.frame( date = seq(as.Date("1910/1/1"), as.Date("1999/1/1"), "years"), x = seq(1:90))
reindex(x$x, time =  x$date, base = "1985-01-01")



}
