% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.R
\name{variation}
\alias{variation}
\title{Title}
\usage{
variation(x, n, time, percent = F)
}
\arguments{
\item{x}{input data set}

\item{n}{a positive integer of length 1 or period given by lubridate package. See statar and lubridate packages}

\item{time}{time variable}

\item{percent}{if true, the result is multiplied by 100}
}
\description{
Title
}
\examples{
library(lubridate)
x <- data.frame(date = as.Date(c('2000-01-01', '2000-02-01', '2000-03-01')), x = 1:3)
variation(x$x, n = months(1), time =  x$date)

}
