% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twelve_months_cumulative.R
\name{twelve_months_cumulative}
\alias{twelve_months_cumulative}
\title{twelve_months_cumulative}
\usage{
twelve_months_cumulative(x, time, percent = F)
}
\arguments{
\item{percent}{}
}
\value{

}
\description{
twelve_months_cumulative
}
\examples{
library(lubridate)
x <- data.frame( date = seq(as.Date("1910/1/1"), as.Date("1999/1/1"), "years"), x = seq(1:90))
twelve_months_cumulative(x$x, time =  x$date)


}
