% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.R
\name{variation}
\alias{variation}
\title{Variation}
\usage{
variation(x, n, time, percent = F)
}
\arguments{
\item{x}{input data set}

\item{n}{a positive integer of length 1 or period given by lubridate package. See statar and lubridate packages}

\item{time}{time variable}

\item{percent}{if true, the result is multiplied by 100}
}
\description{
Variation
}
\examples{
library(lubridate)
x <- data.frame( date = seq(as.Date("1910/1/1"), as.Date("1999/1/1"), "years"), x = seq(1:90))
variation(x$x, n = years(1), time =  x$date)

}
